+++ client/src/services/ApiClient.js
@@ -18,3 +18,84 @@ export async function getEmployeeByID(employeeID) {
       throw new Error("An error occurred during employee fetch");
   }
 }
+
+export async function getOrders() {
+  // Returns mock data right now
+
+  const mockOrders = [
+    {
+      id: "618f10ce364006b8655df032",
+      clientId: "918d971d89d6240eb03742d7",
+      products: [
+        { productId: "718d971d89d6240eb03742d7", quantity: 10 },
+        { productId: "298d971d89d6240eb03742d7", quantity: 1 },
+        { productId: "318d971d89d6240eb03742d7", quantity: 2 },
+        { productId: "418d971d89d6240eb03742d7", quantity: 3 },
+      ],
+      status: "PREPARED",
+      totalPrice: 31,
+      location: "corso Duca degli Abruzzi, 129, Torino",
+      createdAt:
+        new Date().getUTCMonth() +
+        1 +
+        "-" +
+        new Date().getUTCDate() +
+        "-" +
+        new Date().getUTCFullYear(),
+    },
+
+    {
+      id: "718f10ce364006b8655df032",
+      clientId: "918d971d89d6240eb03742d7",
+      products: [
+        { productId: "718d971d89d6240eb03742d7", quantity: 10 },
+        { productId: "298d971d89d6240eb03742d7", quantity: 1 },
+        { productId: "318d971d89d6240eb03742d7", quantity: 2 },
+        { productId: "418d971d89d6240eb03742d7", quantity: 3 },
+      ],
+      status: "DONE",
+      totalPrice: 14,
+      location: "corso Duca degli Abruzzi, 129, Torino",
+      createdAt:
+        new Date().getUTCMonth() +
+        1 +
+        "-" +
+        new Date().getUTCDate() +
+        "-" +
+        new Date().getUTCFullYear(),
+    },
+
+    {
+      id: "818f10ce364006b8655df032",
+      clientId: "918d971d89d6240eb03742d7",
+      products: [{ productId: "718d971d89d6240eb03742d7", quantity: 3 }],
+      status: "PREPARED",
+      totalPrice: 7,
+      location: "corso Duca degli Abruzzi, 129, Torino",
+      createdAt:
+        new Date().getUTCMonth() +
+        1 +
+        "-" +
+        new Date().getUTCDate() +
+        "-" +
+        new Date().getUTCFullYear(),
+    },
+    {
+      id: "118f10ce364006b8655df032",
+      clientId: "918d971d89d6240eb03742d7",
+      products: [{ productId: "118d971d89d6240eb03742d7", quantity: 3 }],
+      status: "PREPARED",
+      totalPrice: 15,
+      location: "corso Duca degli Abruzzi, 129, Torino",
+      createdAt:
+        new Date().getUTCMonth() +
+        1 +
+        "-" +
+        new Date().getUTCDate() +
+        "-" +
+        new Date().getUTCFullYear(),
+    },
+  ];
+
+  return mockOrders;
+}
